requirements.txt
streamlit
langchain-groq
langchain-core
pyyaml